accessible/tests/mochitest/value/test_progress.html

<html>

<head>
  <title>nsIAccessible value testing for progress element</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../value.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/chrome-harness.js"></script>

  <script type="application/javascript">
    function doTest() {
      // HTML5 progress element tests
      testValue("pr_indeterminate", "", 0, 0, 1, 0);
      testValue("pr_zero", "0%", 0, 0, 1, 0);
      testValue("pr_zeropointfive", "50%", 0.5, 0, 1, 0);
      testValue("pr_one", "100%", 1, 0, 1, 0);
      testValue("pr_42", "100%", 42, 0, 1, 0);
      testValue("pr_21", "50%", 21, 0, 42, 0);
      testValue("pr_valuetext", "value", 0, 0, 1, 0);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>

</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=559773"
     title="make HTML5 progress element accessible">
    Mozilla Bug 559773
  </a><br />
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>
  <pre id="test">
  </pre>

  <!-- HTML5 progress element -->
  <progress id="pr_indeterminate">this will be read by legacy browsers</progress>
  <progress id="pr_zero" value="0">this will be read by legacy browsers</progress>
  <progress id="pr_zeropointfive" value="0.5">this will be read by legacy browsers</progress>
  <progress id="pr_one" value="1">this will be read by legacy browsers</progress>
  <progress id="pr_42" value="42">this will be read by legacy browsers</progress>
  <progress id="pr_21" value="21" max="42">this will be read by legacy browsers</progress>
  <!-- aria-valuetext should work due to implicit progressbar role (bug 1475376) -->
  <progress id="pr_valuetext" aria-valuetext="value">this will be read by legacy browsers</progress>
</body>
</html>
